base/requirements/base.txt
# ------------------------------------------------------------------------------
# Django
# ------------------------------------------------------------------------------
# https://www.djangoproject.com/
django==4.0.9  # pyup: < 4.1
# https://github.com/joke2k/django-environ
django-environ==0.9.0
# https://github.com/jazzband/django-model-utils
django-model-utils==4.3.1
# https://github.com/pennersr/django-allauth
django-allauth==0.52.0
# https://github.com/django-crispy-forms/django-crispy-forms
django-crispy-forms==1.14.0
# https://github.com/django-crispy-forms/crispy-bootstrap5
crispy-bootstrap5==0.7
# https://github.com/django-compressor/django-compressor
django-compressor==4.3.1
# https://github.com/adamchainz/django-cors-headers
django-cors-headers==3.13.0
# https://github.com/django-webpack/django-webpack-loader
django-webpack-loader==1.8.1
# For smooth handling of static files.
# https://github.com/evansd/whitenoise
whitenoise==6.2.0
# ------------------------------------------------------------------------------
# Oauth2
# ------------------------------------------------------------------------------
requests
requests-oauthlib
# https://github.com/django-otp/django-otp
django-otp==1.1.4
# https://django-formtools.readthedocs.io/en/latest/
django-formtools==2.4
# https://django-allauth-2fa.readthedocs.io/en/latest/
django-allauth-2fa==0.10.0
# ------------------------------------------------------------------------------
# Extras
# ------------------------------------------------------------------------------
# https://github.com/tqdm/tqdm
tqdm==4.64.1
